pyproject.toml
[project]
name = "hazmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncer>=0.0.8",
    "flask>=3.1.1",
    "frozendict>=2.4.6",
    "openai>=1.0.0",
    "pydantic-ai>=0.0.18",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "requests-cache>=1.2.1",
    "requests-oauthlib>=2.0.0",
    "rich>=14.0.0",
    "typer>=0.16.0",
    "yarl>=1.20.1",
    "loguru>=0.7.3",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "tiktoken>=0.9.0",
    "langchain-google-genai>=2.1.5",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "types-pyyaml>=6.0.12.20250516",
    "types-requests-oauthlib>=2.0.0.20250516",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
